+++ NEWS.md
@@ -1,3 +1,8 @@
+# spex dev
+
+* `polygonize`, `qm_rasterToPolygons`, and `qm_rasterToPolygons_sp` gain `na.rm` 
+ argument to optionally remove any cells that are `NA` (in the first layer)
+
 # spex 0.3.0
 
 * new function `buffer_extent` for whole-grain expansion